feat: updated styles
