+++ ui-components/src/blocks/StandardCard.stories.tsx
@@ -3,11 +3,10 @@ import React from "react"
 import { StandardCard } from "./StandardCard"
 import { BADGES } from "../../.storybook/constants"
 import { MinimalTable } from "../tables/MinimalTable"
+import { mockDataWithStyling, mockHeadersWithStyling } from "../tables/MinimalTable.stories"
+
 import { Button } from "../actions/Button"
 import StandardCardDocumentation from "./StandardCard.docs.mdx"
-import { TableHeaders, StandardTableData } from "../tables/StandardTable"
-import Icon, { IconFillColors } from "../icons/Icon"
-import { faClone, faPenToSquare, faTrashCan } from "@fortawesome/free-regular-svg-icons"
 
 export default {
   title: "Blocks/StandardCard  🚩",
@@ -30,52 +29,6 @@ export const Blank = () => (
   />
 )
 
-const mockHeadersWithStyling: TableHeaders = {
-  name: { name: "t.name", className: "pl-8" },
-  relationship: { name: "t.relationship", className: "pl-8" },
-  dob: {
-    name: "application.household.member.dateOfBirth",
-    className: "pl-8",
-  },
-  icons: { name: "", className: "pl-8" },
-}
-
-const iconContent = () => {
-  return (
-    <div className={"text-right mr-4 w-max"}>
-      <Icon
-        symbol={faPenToSquare}
-        size={"medium"}
-        fill={IconFillColors.primary}
-        className={"mr-5"}
-      />
-      <Icon symbol={faClone} size={"medium"} fill={IconFillColors.primary} className={"mr-5"} />
-      <Icon symbol={faTrashCan} size={"medium"} fill={IconFillColors.alert} />
-    </div>
-  )
-}
-const mockDataWithStyling: StandardTableData = [
-  {
-    name: { content: "Jim Halpert" },
-    relationship: { content: "Husband" },
-    dob: { content: "05/01/1985" },
-    icons: { content: iconContent() },
-  },
-  {
-    name: { content: "Michael Scott" },
-    relationship: { content: "Friend" },
-    dob: { content: "05/01/1975" },
-    icons: { content: iconContent() },
-  },
-]
-
-let i = 5
-while (i > 0) {
-  mockDataWithStyling.push(mockDataWithStyling[0])
-  mockDataWithStyling.push(mockDataWithStyling[1])
-  i--
-}
-
 export const WithTable = () => (
   <StandardCard
     title="Standard Card Title"
@@ -83,12 +36,7 @@ export const WithTable = () => (
     footer={<Button>Add item</Button>}
   >
     <div>
-      <MinimalTable
-        headers={mockHeadersWithStyling}
-        data={mockDataWithStyling}
-        cellClassName={"py-2"}
-        draggable={true}
-      />
+      <MinimalTable headers={mockHeadersWithStyling} data={mockDataWithStyling} draggable={true} />
     </div>
   </StandardCard>
 )
