+++ ui-components/src/locales/general.json
@@ -883,6 +883,7 @@
   "t.occupancy": "Occupancy",
   "t.ok": "Ok",
   "t.or": "or",
+  "t.order": "Order",
   "t.people": "people",
   "t.perMonth": "per month",
   "t.perYear": "per year",
